pinephone: move volume control to BootSequence
This follows advice in
https://www.alsa-project.org/alsa-doc/alsa-lib/group__ucm__conf.html :

"It is expected that the applications handle the volume settings. It is not recommended to set the fixed values for the volume settings to the Enable / Disable sequences for verbs or devices, if the device exports the hardware volume (MixerElem or Volume/Switch values). The default volume settings should be set in the BootSequence."